Documents/code_projects/agent_inbox_1.13/requirements.txt
# Essential dependencies for MVP
python-dotenv
pydantic

# LangChain & LangGraph (latest stable versions)
langchain
langchain-openai
langchain-anthropic
langgraph
langgraph-cli
langsmith
langgraph-api
langgraph-sdk
langchain-mcp-adapters


# CLI
typer[all]
rich

# HTTP client for LangGraph API
httpx

# Logging
structlog

# Google Workspace API dependencies
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
